requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
pillow==11.3.0
pywebview==4.4.2
PyMySQL==1.1.2
python-dotenv==1.1.1
requests==2.32.3
reportlab==4.4.4
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
